travel-butler/requirements_gemini.txt
# Core dependencies for Gemini integration
google-generativeai>=0.8.0  # Gemini SDK
fastapi>=0.109.0            # Web framework
uvicorn[standard]>=0.27.0   # ASGI server
pydantic>=2.6.0             # Data validation
python-dotenv>=1.0.0        # Environment variables

# For async support
httpx>=0.26.0               # Async HTTP client
aiofiles>=23.2.0            # Async file operations

# Optional but recommended
pydantic-settings>=2.1.0    # Settings management